spliting words in string for comparision
Previous solution was simply printing entire paragraphs without itterating through each word. Assigned paragraph to variable and used for loop to split the string into a list of words. Compared each word of all paragraphs until a match was found. Printed the entire paragraph and exited the program since only the very first paragraph with the word in it had to be printed.